OBPIH-7560 Corrections after copilot code review
