+++ src/js/components/Layout/v2/Section.jsx
@@ -6,11 +6,15 @@ import Translate from 'utils/Translate';
 
 import './styles.scss';
 
-const Section = ({ title, children, className }) => (
+const Section = ({
+  showTitle, title, children, className,
+}) => (
   <div className={`v2-section ${className}`}>
+    {showTitle && (
     <span className="v2-section-title text-uppercase">
       <Translate id={title?.label} defaultMessage={title?.defaultMessage} />
     </span>
+    )}
     {children}
   </div>
 );
@@ -18,14 +22,20 @@ const Section = ({ title, children, className }) => (
 export default Section;
 
 Section.propTypes = {
+  showTitle: PropTypes.bool,
   title: PropTypes.shape({
     label: PropTypes.string.isRequired,
     defaultMessage: PropTypes.string.isRequired,
-  }).isRequired,
+  }),
   children: PropTypes.node.isRequired,
   className: PropTypes.string,
 };
 
 Section.defaultProps = {
+  showTitle: true,
+  title: {
+    label: '',
+    defaultMessage: '',
+  },
   className: '',
 };
