+++ src/js/components/stock-movement-wizard/inboundV2/sections/send/InboundSend.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 
 import PropTypes from 'prop-types';
 
-import InboundSendForm from 'components/stock-movement-wizard/inboundV2/sections/InboundSendForm';
-import InboundSendTable from 'components/stock-movement-wizard/inboundV2/sections/InboundSendTable';
+import InboundSendForm from 'components/stock-movement-wizard/inboundV2/sections/send/InboundSendForm';
+import InboundSendTable from 'components/stock-movement-wizard/inboundV2/sections/send/InboundSendTable';
 
 const InboundSend = ({ previous }) => (
   <div className="inbound-send-shipment">
